feat: auto complete 0 value without unit in animation style
When the user specifies aniimation style like `[100px, 0]`, it did not work in the prevoius implementation because `to` value does not have a unit and css transition expression is constructed based on it. By this commit, if to style value is 0 without unit, it automatically complete the unit from `from` style. So the above style will be treated as `[100px, 0px]`